(PC-4939) Make title icon page reusable
